refactor: Add a base DataView class
I created a base `DataView` class to be extended by other classes like
`TableView` and `GridView`, this base class has all the common features
