+++ src/Views/GridView.php
@@ -2,7 +2,7 @@
 
 namespace LaravelViews\Views;
 
-class GridView extends TableView
+class GridView extends DataView
 {
     /** Component name */
     protected $view = 'grid-view.grid-view';
@@ -14,13 +14,4 @@ class GridView extends TableView
 
     /** Max cols to be render on xl */
     public $maxCols = 5;
-
-    /**
-     * Sets the default headers, which are needed on the TableView,
-     * it is an empty array since those aren't needed for a grid view
-     */
-    protected function headers()
-    {
-        return [];
-    }
 }
