+++ tests/TestCase.php
@@ -5,7 +5,7 @@ namespace LaravelViews\Test;
 use LaravelViews\LaravelViewsServiceProvider;
 use Livewire\LivewireServiceProvider;
 use Orchestra\Testbench\TestCase as TestbenchTestCase;
-use PHPUnit\Framework\TestCase as FrameworkTestCase;
+use Spatie\LaravelRay\RayServiceProvider;
 
 class TestCase extends TestbenchTestCase
 {
@@ -26,7 +26,8 @@ class TestCase extends TestbenchTestCase
     {
         return [
             LivewireServiceProvider::class,
-            LaravelViewsServiceProvider::class
+            LaravelViewsServiceProvider::class,
+            RayServiceProvider::class
         ];
     }
 }
